resolve  reviewer  comments
