+++ ImageFeed/ImageListView/ImagesListPresenter.swift
@@ -6,7 +6,6 @@
 //
 
 import Foundation
-import UIKit
 
 final class ImagesListPresenter: ImagesListPresenterProtocol {
